+++ Tests/MapboxMapsTests/Annotations/Mocks/ViewAnnotationUpdateObserver.swift
@@ -1,0 +1,15 @@
+import Foundation
+@testable import MapboxMaps
+
+final class MockViewAnnotationUpdateObserver: ViewAnnotationUpdateObserver {
+
+    let framesDidChangeStub = Stub<[UIView], Void>()
+    func framesDidChange(for annotationViews: [UIView]) {
+        framesDidChangeStub.call(with: annotationViews)
+    }
+
+    let visibilityDidChangeStub = Stub<[UIView], Void>()
+    func visibilityDidChange(for annotationViews: [UIView]) {
+        visibilityDidChangeStub.call(with: annotationViews)
+    }
+}